perf: put parsedPath and pointWidth into context, so multiple components don't have to calculate it by themselves
