+++ src/charts/line/Dot.tsx
@@ -8,7 +8,7 @@ import Animated, {
   withTiming,
 } from 'react-native-reanimated';
 import { Circle, CircleProps } from 'react-native-svg';
-import { getYForX, parse } from 'react-native-redash';
+import { getYForX } from 'react-native-redash';
 
 import { LineChartDimensionsContext } from './Chart';
 import { LineChartPathContext } from './ChartPath';
@@ -57,8 +57,8 @@ export function LineChartDot({
   size = 4,
   outerSize = size * 4,
 }: LineChartDotProps) {
-  const { data, isActive } = useLineChart();
-  const { path, pathWidth: width } = React.useContext(
+  const { isActive } = useLineChart();
+  const { parsedPath, pointWidth } = React.useContext(
     LineChartDimensionsContext
   );
 
@@ -69,17 +69,6 @@ export function LineChartDot({
   const color = isInactive ? inactiveColor || defaultColor : defaultColor;
   const opacity = isInactive && !inactiveColor ? 0.5 : 1;
   const hasOuterDot = defaultHasOuterDot || hasPulse;
-
-  ////////////////////////////////////////////////////////////
-
-  const parsedPath = React.useMemo(() => parse(path), [path]);
-
-  ////////////////////////////////////////////////////////////
-
-  const pointWidth = React.useMemo(
-    () => width / (data.length - 1),
-    [data.length, width]
-  );
 
   ////////////////////////////////////////////////////////////
